[Native File System] Create FileSystemFileWriterImpl to mirror the renderer side
This is the initial cl of the refactor of NativeFileSystemFileHandleImpl to
mirror the split on the renderer side with the handle and the writer as
two separate entities.

It is the initial implementation for FileSystemFileWriterImpl. It includes
only one no-op method for now: close().

This sets the stage for atomic writes as close() will be the explicit
termination of a write, which allows us to do the atomic operation.

Bug: 968550
Change-Id: I0af81efa46789f18c4a8b0bca3ab89c5703b798c